requirements.txt
asgi-lifespan==2.1.0
pydantic==2.12.3
pydantic[email]==2.12.3
pydantic-settings==2.11.0
python-dotenv==1.1.1
redis==7.0.0
sqlalchemy==2.0.44
asyncpg==0.30.0
pytest==8.4.2
pytest-asyncio==1.2.0
uvicorn==0.38.0
fastapi==0.120.0
httpx==0.28.1
alembic==1.17.0
psycopg2-binary==2.9.11
pwdlib[argon2]==0.3.0
loguru==0.7.3
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
requests==2.32.5
pytest-mock==3.15.1
#uvloop